move learnroom entities to shared path
