--- apps/server/src/entities/index.ts
@@ -1,3 +1,0 @@
-export * from './user.entity';
-
-export * from './learnroom';
