--- apps/server/src/entities/learnroom/index.ts
@@ -1,4 +1,0 @@
-export * from './course.entity';
-// export * from './lesson.entity';
-export * from './coursegroup.entity';
-export * from './testHelper';
