+++ apps/server/src/modules/lernroom/mapper/dashboard.mapper.ts
@@ -1,4 +1,4 @@
-import { DashboardEntity } from '../../../entities/learnroom/dashboard.entity';
+import { DashboardEntity } from '../../../shared/domain/entity/dashboard.entity';
 import { DashboardResponse, DashboardGridElementResponse } from '../controller/dto';
 
 export class DashboardMapper {
