+++ apps/server/src/modules/task/repo/submission.repo.integration.spec.ts
@@ -1,7 +1,7 @@
 import { EntityManager, ObjectId } from '@mikro-orm/mongodb';
 import { Test, TestingModule } from '@nestjs/testing';
 
-import { Course } from '@src/entities';
+import { Course } from '@shared/domain';
 import { MongoMemoryDatabaseModule } from '../../database';
 import { FileTaskInfo, LessonTaskInfo, Submission, Task, UserTaskInfo, CourseGroupInfo } from '../entity';
