+++ apps/server/src/shared/domain/entity/course.entity.spec.ts
