+++ apps/server/src/shared/domain/entity/coursegroup.entity.spec.ts
