+++ apps/server/src/shared/domain/entity/index.ts
@@ -1,1 +1,4 @@
 export * from './base.entity';
+export * from './course.entity';
+export * from './coursegroup.entity';
+export * from './dashboard.entity';
