+++ apps/server/src/shared/domain/entity/user.entity.ts
